backend/requirements.txt
# File: backend/.env
fastapi
uvicorn
python-dotenv
google-generativeai